Create StdDraw.java
Standard drawing library. This class provides a basic capability for creating drawings with your programs. It uses a simple graphics model that allows you to create drawings consisting of geometric shapes (e.g., points, lines, circles, rectangles) in a window on your computer and to save the drawings to a file.